requirements.txt
Pandas 
MySQL-connector-python
Streamlit
